requirements.txt
streamlit 
python-dotenv 
crewai 
crewai-tools
pysqlite3